+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 FEATURES:
 * **New Resource:** `vcd_nsxv_dnat` DNAT for advanced edge gateways using proxied NSX API - [#xxx]
-* **New Resource:** `vcd_nsxv_snat` 
+* **New Resource:** `vcd_nsxv_snat`  SNAT for advanced edge gateways using proxied NSX API - [#xxx]
 * **New Data Source:** `vcd_org` Organization - ([#218])
 * **New Data Source:** `vcd_catalog` Catalog - ([#218])
 * **New Data Source:** `vcd_catalog_item` CatalogItem - ([#218])